+++ src/test/java/wooteco/subway/controller/StationAcceptanceTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.station;
+package wooteco.subway.controller;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
